Refactor: extract CellsSweeper
