+++ atlasdb-dagger/src/main/java/com/palantir/atlasdb/services/SweeperModule.java
@@ -21,6 +21,7 @@ import javax.inject.Singleton;
 import com.google.common.collect.ImmutableList;
 import com.palantir.atlasdb.cleaner.Follower;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
+import com.palantir.atlasdb.sweep.CellsSweeper;
 import com.palantir.atlasdb.sweep.SweepTaskRunner;
 import com.palantir.atlasdb.sweep.SweepTaskRunnerImpl;
 import com.palantir.atlasdb.transaction.impl.SerializableTransactionManager;
@@ -41,13 +42,12 @@ public class SweeperModule {
                                                   SweepStrategyManager sweepStrategyManager,
                                                   Follower follower) {
         return new SweepTaskRunnerImpl(
-                txm,
                 kvs,
                 txm::getUnreadableTimestamp,
                 txm::getImmutableTimestamp,
                 transactionService,
                 sweepStrategyManager,
-                ImmutableList.of(follower));
+                new CellsSweeper(txm, kvs, ImmutableList.of(follower)));
     }
 
 }
