+++ atlasdb-dagger/src/main/java/com/palantir/atlasdb/services/test/TestSweeperModule.java
@@ -23,6 +23,7 @@ import com.google.common.base.Supplier;
 import com.google.common.collect.ImmutableList;
 import com.palantir.atlasdb.cleaner.Follower;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
+import com.palantir.atlasdb.sweep.CellsSweeper;
 import com.palantir.atlasdb.sweep.SweepTaskRunner;
 import com.palantir.atlasdb.sweep.SweepTaskRunnerImpl;
 import com.palantir.atlasdb.transaction.impl.SerializableTransactionManager;
@@ -66,13 +67,12 @@ public class TestSweeperModule {
         Supplier<Long> unreadable = unreadableTs.isPresent() ? unreadableTs.get() : txm::getUnreadableTimestamp;
         Supplier<Long> immutable = immutableTs.isPresent() ? immutableTs.get() : txm::getImmutableTimestamp;
         return new SweepTaskRunnerImpl(
-                txm,
                 kvs,
                 unreadable,
                 immutable,
                 transactionService,
                 sweepStrategyManager,
-                ImmutableList.of(follower));
+                new CellsSweeper(txm, kvs, ImmutableList.of(follower)));
     }
 
 }
